requirements.txt
fastapi==0.74.1
pyjwt==2.8.0
python-decouple==3.8
uvicorn==0.17.6
sqlalchemy==1.4.25
pydantic==1.8.2